requirements.txt
requests
langchain == 0.3.7
google-generativeai == 0.8.3
langchain-community == 0.3.7
langchainhub == 0.1.21
python-dotenv
langchain_google_genai
flask
